+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Piece.trim/1` to remove starting square from a three element tuple containing
+  a piece name, color, and starting square.
+
 ### Changed
 
 - `Parser.FEN.parse/1` and `Parser.FEN.serialize/1` now return a tuple with